Compiling again, but need to update print_lir.c

+++ src/nibble.h
@@ -75,7 +75,7 @@ typedef enum FloatKind {
     FLOAT_KIND_COUNT
 } FloatKind;
 
-extern const size_t float_kind_sizes[FLOAT_KIND_COUNT];
+extern const u8 float_kind_sizes[FLOAT_KIND_COUNT];
 extern const char* float_kind_names[FLOAT_KIND_COUNT];
 
 typedef struct Float {
@@ -111,7 +111,7 @@ typedef enum IntegerKind {
     INTEGER_KIND_COUNT
 } IntegerKind;
 
-extern const size_t int_kind_sizes[INTEGER_KIND_COUNT];
+extern const u8 int_kind_sizes[INTEGER_KIND_COUNT];
 extern const char* int_kind_names[INTEGER_KIND_COUNT];
 extern const bool int_kind_signed[INTEGER_KIND_COUNT];
 extern const u64 int_kind_max[INTEGER_KIND_COUNT];
